Extract booking payment policy runtime
